refactor(app): split startup app responsibilities
